+++ src/dialoged/skein/index.ts
@@ -13,5 +13,6 @@ export * from './io';
 export * from './project';
 export * from './progress';
 export * from './trace';
+export * from './frotz-build';
 export * from './syntax';
 export * from './search';